requirements.txt
langchain==0.2.16
langchain-openai==0.1.25
langgraph==0.2.28
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2
